+++ graphql-maven-plugin-logic/src/main/resources/templates/server_GraphQLProvider.vm.java
@@ -161,7 +161,9 @@ public class GraphQLProvider {
 			// Let's link the interface types to the concrete types
 #end
 #foreach ($interface in $interfaces)
+#if ($interface.implementingTypes.size() >0)
 			.type("${interface.javaName}", typeWiring -> typeWiring.typeResolver(get${interface.javaName}Resolver()))
+#end
 #end
 #if ($unions.size() > 0)
 			//
@@ -174,6 +176,7 @@ public class GraphQLProvider {
 	}
 
 #foreach ($interface in $interfaces)
+#if ($interface.implementingTypes.size() >0)
 	private TypeResolver get${interface.javaName}Resolver() {
 		return new TypeResolver() {
 			@Override
@@ -195,6 +198,7 @@ public class GraphQLProvider {
 		};
 	}
 
+#end
 #end
 #foreach ($union in $unions)
 private TypeResolver get${union.javaName}Resolver() {